keep trying to whack-a-mole broken e2e tests
